bmqa/requirements.txt
python-telegram-bot==21.0.1
redis==5.0.1
kvsqlite==0.1.3
requests==2.31.0
yt-dlp==2024.1.1
pydub==0.25.1
mutagen==1.46.0
SpeechRecognition==3.10.0
gTTS==2.4.0
Pillow==10.1.0
python-dotenv==1.0.0
pytz==2024.1
akinator==2.0.2

# إضافات المكتبات اللي يستخدمها الكود أو متعلقة بالميزات الاختيارية
psutil==7.2.2
py-cpuinfo
pytio==0.3.1
meval==2.5
httpx==0.28.1
pytube==15.0.0
youtube-search==2.2.0
shazamio==0.8.1

